requirements.txt
numpy
pandas
matplotlib
seaborn
streamlit
watchdog
streamlit-on-Hover-tabs
# datetime
# time
# requests
# os
python-dotenv
streamlit-authenticator
yaml
st_pages
streamlit_extras
